+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 next-version
 ============
 
+0.11.0
+======
+
 - web: improved similarity page render.
 - web: add top 100 least common anomalies section in similarity report.
 - web: use a mono space font for log lines.